refactor Spotify, MusixMatch and Google sign in into /external (#226)
* refactor MusixmatchLyricsGetter.kt into /external/musixmatch

* refactor ImportSpotifyPlaylistsActivity.kt & SpotifyReceiveActivity.kt into /external/spotify

* refactor GoogleSignInActivity.kt into /external/google
