+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockSongFragment.kt
@@ -1,10 +1,10 @@
 package com.github.fribourgsdp.radio.mockimplementations
 
-import com.github.fribourgsdp.radio.LyricsGetter
+import com.github.fribourgsdp.radio.external.musixmatch.LyricsGetter
 import com.github.fribourgsdp.radio.data.view.SongFragment
 
 class MockSongFragment : SongFragment() {
-    override fun getLyricsGetter() : LyricsGetter{
+    override fun getLyricsGetter() : LyricsGetter {
         return MockLyricsGetter
     }
 }
