+++ app/src/main/res/layout/activity_spotify_receive.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SpotifyReceiveActivity">
+    tools:context=".external.spotify.SpotifyReceiveActivity">
 
     <Button
         android:id="@+id/importSpotifyPlaylistsButton"
